pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.31.0",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
